feat: Solución de errores
