+++ Saca la Bici/Data/Models/Network/RouteAPIService.swift
@@ -1,25 +1,10 @@
 import Foundation
 import CoreLocation
 
-struct RouteDetails {
-    let titulo: String
-    let distancia: String
-    let tiempo: String
-    let nivel: String
-    let start: CLLocationCoordinate2D
-    let stopover1: CLLocationCoordinate2D
-    let stopover2: CLLocationCoordinate2D
-    let descanso: CLLocationCoordinate2D
-    let stopover3: CLLocationCoordinate2D
-    let stopover4: CLLocationCoordinate2D
-    let end: CLLocationCoordinate2D
-}
-
 class RouteAPIService {
     static let shared = RouteAPIService()
     let baseURL = "http://18.220.205.53:8080/mapa/registrarRuta"
 
-    // Enviar la ruta al servidor
     func sendRoute(routeDetails: RouteDetails, idToken: String, completion: @escaping (Bool) -> Void) {
         guard let url = URL(string: baseURL) else {
             print("Error: URL inválida")
@@ -27,28 +12,12 @@ class RouteAPIService {
             return
         }
 
-        // Validar coordenadas y distancia
-        if routeDetails.distancia.isEmpty || routeDetails.distancia == "NaN" ||
-           routeDetails.start.latitude.isNaN || routeDetails.start.longitude.isNaN ||
-           routeDetails.stopover1.latitude.isNaN || routeDetails.stopover1.longitude.isNaN ||
-           routeDetails.stopover2.latitude.isNaN || routeDetails.stopover2.longitude.isNaN ||
-           routeDetails.descanso.latitude.isNaN || routeDetails.descanso.longitude.isNaN ||
-           routeDetails.stopover3.latitude.isNaN || routeDetails.stopover3.longitude.isNaN ||
-           routeDetails.stopover4.latitude.isNaN || routeDetails.stopover4.longitude.isNaN ||
-           routeDetails.end.latitude.isNaN || routeDetails.end.longitude.isNaN {
-            print("Error: Coordenadas o distancia inválidas (NaN)")
-            completion(false)
-            return
-        }
-
-        // Solicitud
         var request = URLRequest(url: url)
         request.httpMethod = "POST"
         request.setValue("application/json", forHTTPHeaderField: "Content-Type")
-        request.setValue("Bearer \(idToken)", forHTTPHeaderField: "Authorization")  
+        request.setValue("Bearer \(idToken)", forHTTPHeaderField: "Authorization")
         request.timeoutInterval = 60.0
 
-        // JSON con los datos de la ruta
         let json: [String: Any] = [
             "titulo": routeDetails.titulo,
             "distancia": routeDetails.distancia,
@@ -74,7 +43,7 @@ class RouteAPIService {
             return
         }
 
-        let task = URLSession.shared.dataTask(with: request) { data, response, error in
+        let task = URLSession.shared.dataTask(with: request) { _, response, error in
             if let error = error {
                 print("Error en la solicitud HTTP: \(error.localizedDescription)")
                 completion(false)
@@ -82,21 +51,12 @@ class RouteAPIService {
             }
 
             if let httpResponse = response as? HTTPURLResponse {
-                print("Código de respuesta: \(httpResponse.statusCode)")
-
-                if let data = data, let responseString = String(data: data, encoding: .utf8) {
-                    print("Respuesta del servidor: \(responseString)")
-                }
-
                 if httpResponse.statusCode == 200 {
-                    print("Ruta registrada exitosamente.")
                     completion(true)
                 } else {
-                    print("Error en la respuesta del servidor: Código \(httpResponse.statusCode)")
                     completion(false)
                 }
             } else {
-                print("No se recibió una respuesta HTTP válida.")
                 completion(false)
             }
         }
