+++ Saca la Bici/Framework/Views/Mapa/MapaView.swift
@@ -6,13 +6,13 @@ struct MapaView: View {
     @State private var routeCoordinates: [CLLocationCoordinate2D] = []
     @State private var distance: Double = 0.0
     @State private var message: String = ""
-
+    
     var body: some View {
         NavigationView {
             ZStack {
                 MapViewContainer(routeCoordinates: $routeCoordinates, distance: $distance, isAddingRoute: $isAddingRoute, message: $message)
                     .ignoresSafeArea()
-
+                
                 VStack {
                     HStack {
                         Spacer()
